Finished ANOVA part of regression module, so that module is now complete. Refactored all code to comply with PEP8 style guide.
In utils module, removed checkValidCategoricals() and checkNumericalVarsRequested(), as well as any code in other modules that uses them to check validity of requested numerical or categorical vars.
This is because the argparse library provides in its add_argument() method a "choices" parameter that allows us to easily specify valid vars, and allows argparse to handle the invalid requests itself.
Added a new function to utils module called getNumericals(data), which returns a list of names of numerical vars/columns in the dataframe. This will be used in conjunction with the aforementioned choices parameter to check for errors.
For categoricals, since any var in the dataframe is a valid categorical, a list of valid categorical var names can be got using data.columns.values. This too will be used in conjunction with the choices parameter.